deta/requirements.txt
aiogram==3.1.1
aiohttp==3.8.6
aiohttp-jinja2==1.5.1
python-dotenv==1.0.0
aiohttp-ratelimiter==4.1.2
